move instances to std module, fix instance names
